deleteSection with hooks

+++ controllers/sectionController.js
@@ -1,4 +1,6 @@
 const Section = require('../models/section');
+const Admin = require('../models/admin');
+const { use } = require('../routes/api');
 
 exports.createSection = async (req, res) => {
     try {
@@ -41,6 +43,19 @@ exports.updateSection = async (req, res) => {
 
 exports.deleteSection = async (req, res) => {
     try {
+        const admin_id = req.user._id;
+        const mot_de_passe = req.body.mot_de_passe;
+        const admin = await Admin.findById(admin_id);
+        console.log(mot_de_passe);
+        if (!mot_de_passe) {
+            return res.status(400).json({ status: false, message: 'Les paramètres ne sont pas corrects' });
+        }
+        const isMot_de_passeCorrect = await admin.compareMot_de_passe(mot_de_passe);
+        if (!isMot_de_passeCorrect) {
+            return res.status(401).json({ status: false, message: 'password incorrect' });
+        }
+
+
         await Section.findByIdAndDelete(req.params.id);
         res.status(204).send();
     } catch (error) {
